Used a different base image for MySQL.
This `mysql/mysql-server` official image from the Oracle MySQL team is smaller and starts faster than the Docker "official" `mysql` image. It's also far simpler to wait for startup to complete (the `mysql` image initializes MySQL, stops it, then restarts it, so the "ready" messages appear twice, making it very difficult for the the docker maven plugin to wait reliably).