DBZ-121 Add flexible topic naming strategy to MySQL connector
The MySQL connector now has a configuration property that controls how changes events for tables are mapped to topics. By default, all change events for each table are written to a separate topic (this was the old behavior). This strategy is enabled with the `table` literal.

It is also possible to have all change events for all tables in a database to be written to one topic named for the database. The change events for each table (version) will have a different schema, but this strategy may be more attractive for some scenarios that need to more easily consume all change events for a transaction. This strategy is enabled with the `database` literal.

Another option is to use a regular expression and a replacement pattern. The regular expression matches the fully-qualified name of a table, and when matched the replacement pattern (which can use the capture groups, such as `$1_$2`) is used to determine the topic name.

If custom logic is needed, the strategy can also specify the fully-qualified name of a class that implements the `io.debezium.relational.topic.TopicNamingStrategy’ interface.

Finally, it is possible to combine multiple strategies by specifying a comma-separated list of strategies. They will be applied in order, and in these cases it is recommended to use either `table` or `database` literal at the end of the list for a fallback.